+++ test/javascript/public/test/settings.js
@@ -3,7 +3,7 @@ QUnit.config.autostart = window.location.search.search('autostart=false') < 0
 QUnit.config.urlConfig.push({
   id: 'jquery',
   label: 'jQuery version',
-  value: ['3.5.0', '3.5.0.slim', '3.4.1', '3.4.1.slim', '3.3.1', '3.3.1.slim', '3.2.1', '3.2.1.slim', '3.1.1', '3.1.1.slim', '3.0.0', '3.0.0.slim', '2.2.4', '2.1.4', '2.0.3', '1.12.4', '1.11.3'],
+  value: ['3.5.1', '3.5.1.slim', '3.4.1', '3.4.1.slim', '3.3.1', '3.3.1.slim', '3.2.1', '3.2.1.slim', '3.1.1', '3.1.1.slim', '3.0.0', '3.0.0.slim', '2.2.4', '2.1.4', '2.0.3', '1.12.4', '1.11.3'],
   tooltip: 'What jQuery Core version to test against'
 })