+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * [BUGFIX] Fix Rails generators ([#786](https://github.com/DavyJonesLocker/client_side_validations/pull/786))
 * [BUGFIX] Do not validate dynamically disabled inputs ([#789](https://github.com/DavyJonesLocker/client_side_validations/pull/789))
 * [BUGFIX] Fix removal of error messages ([#790](https://github.com/DavyJonesLocker/client_side_validations/pull/790))
+* [ENHANCEMENT] Test against jQuery 3.5.1
 
 ## 16.2.0 / 2020-04-10